I Added 'update_user_by_id' command to CLI
A new command, 'update_user_by_id', has been implemented and added to the command line interface. This command allows the updating of a user's details by their ID. The command reads existing user data from a JSON file, searches for a user with the provided ID, and if found, updates the user's details with the provided options. The updated data is then written back to the JSON file.